backend/requirements.txt
Django==3.2.25
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.9
Pillow==10.3.0
PyYAML==6.0.1
gunicorn==21.2.0
